requirements.txt
langchain
fastapi
uvicorn
langchain-community
openai
python-dotenv
langchain_astradb
pandas

-e .